+++ setup.cfg
@@ -42,7 +42,7 @@ per-file-ignores =
   # names for to comply with the openapi convention (as `item`, `info`, etc):
   dmr/openapi/objects/*.py: WPS201, WPS110
   # We have complex modules with a lot of quircks:
-  dmr/endpoint.py: WPS203, WPS402
+  dmr/endpoint.py: WPS201, WPS203, WPS402
   dmr/validation/endpoint_metadata.py: WPS402
   dmr/openapi/mappers/schema_loader.py: WPS202
   dmr/controller.py: WPS201
